remove wallet reference from servlet. Move transaction related function from wallet to TransactionUtil.
