+++ actuator/src/main/java/org/tron/core/utils/TransactionUtil.java
@@ -15,22 +15,58 @@
 
 package org.tron.core.utils;
 
+import com.google.common.base.CaseFormat;
+import com.google.common.primitives.Longs;
 import com.google.protobuf.Any;
+import com.google.protobuf.ByteString;
 import com.google.protobuf.InvalidProtocolBufferException;
+import java.security.SignatureException;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ArrayUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import org.tron.api.GrpcAPI.Return;
+import org.tron.api.GrpcAPI.Return.response_code;
+import org.tron.api.GrpcAPI.TransactionApprovedList;
+import org.tron.api.GrpcAPI.TransactionExtention;
+import org.tron.api.GrpcAPI.TransactionSignWeight;
+import org.tron.api.GrpcAPI.TransactionSignWeight.Result;
+import org.tron.common.crypto.ECKey;
+import org.tron.common.utils.ByteUtil;
+import org.tron.common.utils.DecodeUtil;
+import org.tron.common.utils.Hash;
 import org.tron.common.utils.Sha256Hash;
+import org.tron.core.ChainBaseManager;
+import org.tron.core.capsule.AccountCapsule;
+import org.tron.core.capsule.ContractCapsule;
+import org.tron.core.capsule.TransactionCapsule;
 import org.tron.core.exception.ContractValidateException;
+import org.tron.core.exception.PermissionException;
+import org.tron.core.exception.SignatureFormatException;
+import org.tron.protos.Protocol.Permission;
+import org.tron.protos.Protocol.Permission.PermissionType;
 import org.tron.protos.Protocol.Transaction;
+import org.tron.protos.Protocol.Transaction.Contract;
 import org.tron.protos.Protocol.Transaction.Contract.ContractType;
 import org.tron.protos.Protocol.Transaction.Result.contractResult;
+import org.tron.protos.Protocol.TransactionSign;
 import org.tron.protos.contract.ShieldContract.ShieldedTransferContract;
 import org.tron.protos.contract.ShieldContract.SpendDescription;
 import org.tron.protos.contract.SmartContractOuterClass.CreateSmartContract;
+import org.tron.protos.contract.SmartContractOuterClass.SmartContract;
+import org.tron.protos.contract.SmartContractOuterClass.SmartContract.ABI;
+import org.tron.protos.contract.SmartContractOuterClass.SmartContract.ABI.Entry.StateMutabilityType;
 import org.tron.protos.contract.SmartContractOuterClass.TriggerSmartContract;
 
 @Slf4j(topic = "capsule")
+@Component
 public class TransactionUtil {
+
+  @Autowired
+  private ChainBaseManager chainBaseManager;
 
   public static boolean validAccountName(byte[] accountName) {
     if (ArrayUtils.isEmpty(accountName)) {
@@ -181,5 +217,184 @@ public class TransactionUtil {
     }
   }
 
+  public static boolean isConstant(SmartContract.ABI abi, byte[] selector) {
+
+    if (selector == null || selector.length != 4
+        || abi.getEntrysList().size() == 0) {
+      return false;
+    }
+
+    for (int i = 0; i < abi.getEntrysCount(); i++) {
+      ABI.Entry entry = abi.getEntrys(i);
+      if (entry.getType() != ABI.Entry.EntryType.Function) {
+        continue;
+      }
+
+      int inputCount = entry.getInputsCount();
+      StringBuilder sb = new StringBuilder();
+      sb.append(entry.getName());
+      sb.append("(");
+      for (int k = 0; k < inputCount; k++) {
+        ABI.Entry.Param param = entry.getInputs(k);
+        sb.append(param.getType());
+        if (k + 1 < inputCount) {
+          sb.append(",");
+        }
+      }
+      sb.append(")");
+
+      byte[] funcSelector = new byte[4];
+      System.arraycopy(Hash.sha3(sb.toString().getBytes()), 0, funcSelector, 0, 4);
+      if (Arrays.equals(funcSelector, selector)) {
+        return entry.getConstant() || entry.getStateMutability().equals(StateMutabilityType.View);
+      }
+    }
+
+    return false;
+  }
+
+  public static byte[] generateContractAddress(Transaction trx) {
+
+    CreateSmartContract contract = ContractCapsule.getSmartContractFromTransaction(trx);
+    byte[] ownerAddress = contract.getOwnerAddress().toByteArray();
+    TransactionCapsule trxCap = new TransactionCapsule(trx);
+    byte[] txRawDataHash = trxCap.getTransactionId().getBytes();
+
+    byte[] combined = new byte[txRawDataHash.length + ownerAddress.length];
+    System.arraycopy(txRawDataHash, 0, combined, 0, txRawDataHash.length);
+    System.arraycopy(ownerAddress, 0, combined, txRawDataHash.length, ownerAddress.length);
+
+    return DecodeUtil.sha3omit12(combined);
+
+  }
+
+  public static byte[] generateContractAddress(byte[] ownerAddress, byte[] txRawDataHash) {
+
+    byte[] combined = new byte[txRawDataHash.length + ownerAddress.length];
+    System.arraycopy(txRawDataHash, 0, combined, 0, txRawDataHash.length);
+    System.arraycopy(ownerAddress, 0, combined, txRawDataHash.length, ownerAddress.length);
+
+    return DecodeUtil.sha3omit12(combined);
+
+  }
+
+  // for `CREATE`
+  public static byte[] generateContractAddress(byte[] transactionRootId, long nonce) {
+    byte[] nonceBytes = Longs.toByteArray(nonce);
+    byte[] combined = new byte[transactionRootId.length + nonceBytes.length];
+    System.arraycopy(transactionRootId, 0, combined, 0, transactionRootId.length);
+    System.arraycopy(nonceBytes, 0, combined, transactionRootId.length, nonceBytes.length);
+
+    return DecodeUtil.sha3omit12(combined);
+  }
+
+  // for `CREATE2`
+  public static byte[] generateContractAddress2(byte[] address, byte[] salt, byte[] code) {
+    byte[] mergedData = ByteUtil.merge(address, salt, Hash.sha3(code));
+    return DecodeUtil.sha3omit12(mergedData);
+  }
+
+  public static boolean checkPermissionOprations(Permission permission, Contract contract)
+      throws PermissionException {
+    ByteString operations = permission.getOperations();
+    if (operations.size() != 32) {
+      throw new PermissionException("operations size must be 32");
+    }
+    int contractType = contract.getTypeValue();
+    boolean b = (operations.byteAt(contractType / 8) & (1 << (contractType % 8))) != 0;
+    return b;
+  }
+
+  public static String makeUpperCamelMethod(String originName) {
+    return "get" + CaseFormat.UPPER_UNDERSCORE.to(CaseFormat.UPPER_CAMEL, originName)
+        .replace("_", "");
+  }
+
+  public static byte[] getSelector(byte[] data) {
+    if (data == null
+        || data.length < 4) {
+      return null;
+    }
+
+    byte[] ret = new byte[4];
+    System.arraycopy(data, 0, ret, 0, 4);
+    return ret;
+  }
+
+  public static TransactionCapsule getTransactionSign(TransactionSign transactionSign) {
+    byte[] privateKey = transactionSign.getPrivateKey().toByteArray();
+    TransactionCapsule trx = new TransactionCapsule(transactionSign.getTransaction());
+    trx.sign(privateKey);
+    return trx;
+  }
+
+  public TransactionCapsule addSign(TransactionSign transactionSign)
+      throws PermissionException, SignatureException, SignatureFormatException {
+    byte[] privateKey = transactionSign.getPrivateKey().toByteArray();
+    TransactionCapsule trx = new TransactionCapsule(transactionSign.getTransaction());
+    trx.addSign(privateKey, chainBaseManager.getAccountStore());
+    return trx;
+  }
+
+  public TransactionSignWeight getTransactionSignWeight(Transaction trx) {
+    TransactionSignWeight.Builder tswBuilder = TransactionSignWeight.newBuilder();
+    TransactionExtention.Builder trxExBuilder = TransactionExtention.newBuilder();
+    trxExBuilder.setTransaction(trx);
+    trxExBuilder.setTxid(ByteString.copyFrom(Sha256Hash.hash(trx.getRawData().toByteArray())));
+    Return.Builder retBuilder = Return.newBuilder();
+    retBuilder.setResult(true).setCode(response_code.SUCCESS);
+    trxExBuilder.setResult(retBuilder);
+    tswBuilder.setTransaction(trxExBuilder);
+    Result.Builder resultBuilder = Result.newBuilder();
+    try {
+      Contract contract = trx.getRawData().getContract(0);
+      byte[] owner = TransactionCapsule.getOwner(contract);
+      AccountCapsule account = chainBaseManager.getAccountStore().get(owner);
+      if (account == null) {
+        throw new PermissionException("Account does not exist!");
+      }
+      int permissionId = contract.getPermissionId();
+      Permission permission = account.getPermissionById(permissionId);
+      if (permission == null) {
+        throw new PermissionException("Permission for this, does not exist!");
+      }
+      if (permissionId != 0) {
+        if (permission.getType() != PermissionType.Active) {
+          throw new PermissionException("Permission type is wrong!");
+        }
+        //check operations
+        if (!checkPermissionOprations(permission, contract)) {
+          throw new PermissionException("Permission denied!");
+        }
+      }
+      tswBuilder.setPermission(permission);
+      if (trx.getSignatureCount() > 0) {
+        List<ByteString> approveList = new ArrayList<ByteString>();
+        long currentWeight = TransactionCapsule.checkWeight(permission, trx.getSignatureList(),
+            Sha256Hash.hash(trx.getRawData().toByteArray()), approveList);
+        tswBuilder.addAllApprovedList(approveList);
+        tswBuilder.setCurrentWeight(currentWeight);
+      }
+      if (tswBuilder.getCurrentWeight() >= permission.getThreshold()) {
+        resultBuilder.setCode(Result.response_code.ENOUGH_PERMISSION);
+      } else {
+        resultBuilder.setCode(Result.response_code.NOT_ENOUGH_PERMISSION);
+      }
+    } catch (SignatureFormatException signEx) {
+      resultBuilder.setCode(Result.response_code.SIGNATURE_FORMAT_ERROR);
+      resultBuilder.setMessage(signEx.getMessage());
+    } catch (SignatureException signEx) {
+      resultBuilder.setCode(Result.response_code.COMPUTE_ADDRESS_ERROR);
+      resultBuilder.setMessage(signEx.getMessage());
+    } catch (PermissionException permEx) {
+      resultBuilder.setCode(Result.response_code.PERMISSION_ERROR);
+      resultBuilder.setMessage(permEx.getMessage());
+    } catch (Exception ex) {
+      resultBuilder.setCode(Result.response_code.OTHER_ERROR);
+      resultBuilder.setMessage(ex.getClass() + " : " + ex.getMessage());
+    }
+    tswBuilder.setResult(resultBuilder);
+    return tswBuilder.build();
+  }
 
 }
