+++ common/src/main/java/org/tron/common/utils/DecodeUtil.java
@@ -2,18 +2,20 @@ package org.tron.common.utils;
 
 import static java.util.Arrays.copyOfRange;
 import static org.tron.common.utils.Hash.sha3;
-import static org.tron.core.Constant.ADD_PRE_FIX_BYTE_MAINNET;
 
 import java.util.Arrays;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ArrayUtils;
 import org.spongycastle.math.ec.ECPoint;
+import org.tron.core.Constant;
 
 @Slf4j(topic = "Commons")
 public class DecodeUtil {
 
   public static final int ADDRESS_SIZE = 42;
-  public static byte addressPreFixByte = ADD_PRE_FIX_BYTE_MAINNET;
+  public static byte addressPreFixByte = Constant.ADD_PRE_FIX_BYTE_MAINNET;
+
+  public static String addressPreFixString = Constant.ADD_PRE_FIX_STRING_MAINNET;
 
   public static boolean addressValid(byte[] address) {
     if (ArrayUtils.isEmpty(address)) {
@@ -32,7 +34,6 @@ public class DecodeUtil {
           + address[0] + " !!");
       return false;
     }
-    //Other rule;
     return true;
   }
