+++ framework/src/main/java/org/tron/core/services/RpcApiService.java
@@ -96,6 +96,7 @@ import org.tron.core.exception.StoreException;
 import org.tron.core.exception.VMIllegalException;
 import org.tron.core.exception.ZksnarkException;
 import org.tron.core.services.ratelimiter.RateLimiterInterceptor;
+import org.tron.core.utils.TransactionUtil;
 import org.tron.core.zen.address.DiversifierT;
 import org.tron.core.zen.address.IncomingViewingKey;
 import org.tron.protos.Protocol;
@@ -159,6 +160,10 @@ public class RpcApiService implements Service {
   private NodeManager nodeManager;
   @Autowired
   private Wallet wallet;
+
+  @Autowired
+  private TransactionUtil transactionUtil;
+
   @Autowired
   private NodeInfoService nodeInfoService;
   @Autowired
@@ -827,7 +832,7 @@ public class RpcApiService implements Service {
     @Override
     public void getTransactionSign(TransactionSign req,
         StreamObserver<Transaction> responseObserver) {
-      TransactionCapsule result = wallet.getTransactionSign(req);
+      TransactionCapsule result = TransactionUtil.getTransactionSign(req);
       responseObserver.onNext(result.getInstance());
       responseObserver.onCompleted();
     }
@@ -838,7 +843,7 @@ public class RpcApiService implements Service {
       TransactionExtention.Builder trxExtBuilder = TransactionExtention.newBuilder();
       Return.Builder retBuilder = Return.newBuilder();
       try {
-        TransactionCapsule trx = wallet.getTransactionSign(req);
+        TransactionCapsule trx = TransactionUtil.getTransactionSign(req);
         trxExtBuilder.setTransaction(trx.getInstance());
         trxExtBuilder.setTxid(trx.getTransactionId().getByteString());
         retBuilder.setResult(true).setCode(response_code.SUCCESS);
@@ -858,7 +863,7 @@ public class RpcApiService implements Service {
       TransactionExtention.Builder trxExtBuilder = TransactionExtention.newBuilder();
       Return.Builder retBuilder = Return.newBuilder();
       try {
-        TransactionCapsule trx = wallet.addSign(req);
+        TransactionCapsule trx = transactionUtil.addSign(req);
         trxExtBuilder.setTransaction(trx.getInstance());
         trxExtBuilder.setTxid(trx.getTransactionId().getByteString());
         retBuilder.setResult(true).setCode(response_code.SUCCESS);
@@ -875,7 +880,7 @@ public class RpcApiService implements Service {
     @Override
     public void getTransactionSignWeight(Transaction req,
         StreamObserver<TransactionSignWeight> responseObserver) {
-      TransactionSignWeight tsw = wallet.getTransactionSignWeight(req);
+      TransactionSignWeight tsw = transactionUtil.getTransactionSignWeight(req);
       responseObserver.onNext(tsw);
       responseObserver.onCompleted();
     }
