+++ framework/src/main/java/org/tron/core/services/http/TransactionSignServlet.java
@@ -5,10 +5,9 @@ import java.util.stream.Collectors;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import org.tron.core.Wallet;
 import org.tron.core.capsule.TransactionCapsule;
+import org.tron.core.utils.TransactionUtil;
 import org.tron.protos.Protocol.Transaction;
 import org.tron.protos.Protocol.TransactionSign;
 
@@ -16,9 +15,6 @@ import org.tron.protos.Protocol.TransactionSign;
 @Component
 @Slf4j(topic = "API")
 public class TransactionSignServlet extends RateLimiterServlet {
-
-  @Autowired
-  private Wallet wallet;
 
   protected void doGet(HttpServletRequest request, HttpServletResponse response) {
 
@@ -38,7 +34,7 @@ public class TransactionSignServlet extends RateLimiterServlet {
       input.put("transaction", jsonTransaction);
       TransactionSign.Builder build = TransactionSign.newBuilder();
       JsonFormat.merge(input.toJSONString(), build, visible);
-      TransactionCapsule reply = wallet.getTransactionSign(build.build());
+      TransactionCapsule reply = TransactionUtil.getTransactionSign(build.build());
       if (reply != null) {
         response.getWriter().println(Util.printCreateTransaction(reply.getInstance(), visible));
       } else {
