+++ framework/src/test/java/org/tron/common/runtime/vm/IsContractTest.java
@@ -12,6 +12,7 @@ import org.tron.core.exception.ContractExeException;
 import org.tron.core.exception.ContractValidateException;
 import org.tron.core.exception.ReceiptCheckErrException;
 import org.tron.core.exception.VMIllegalException;
+import org.tron.core.utils.TransactionUtil;
 import org.tron.core.vm.config.ConfigLoader;
 import org.tron.core.vm.config.VMConfig;
 import org.tron.protos.Protocol.Transaction;
@@ -167,7 +168,7 @@ contract isTestCtr {
     Transaction trx = TvmTestUtils.generateDeploySmartContractAndGetTransaction(
         contractName, address, ABI, factoryCode, value, fee, consumeUserResourcePercent,
         null);
-    byte[] factoryAddress = Wallet.generateContractAddress(trx);
+    byte[] factoryAddress = TransactionUtil.generateContractAddress(trx);
     String factoryAddressStr = Wallet.encode58Check(factoryAddress);
     runtime = TvmTestUtils.processTransactionAndReturnRuntime(trx, rootDeposit, null);
     Assert.assertNull(runtime.getRuntimeError());
@@ -175,7 +176,7 @@ contract isTestCtr {
     trx = TvmTestUtils.generateDeploySmartContractAndGetTransaction(
         "", address, ABI, factoryCode, value, fee, consumeUserResourcePercent,
         null);
-    byte[] factoryAddressOther = Wallet.generateContractAddress(trx);
+    byte[] factoryAddressOther = TransactionUtil.generateContractAddress(trx);
     String factoryAddressStrOther = Wallet.encode58Check(factoryAddressOther);
     runtime = TvmTestUtils.processTransactionAndReturnRuntime(trx, rootDeposit, null);
     Assert.assertNull(runtime.getRuntimeError());
